Remove Rails:: namespace
