+++ README.md
@@ -1,41 +1,11 @@
-# ShiftCommerce::Rails
+This Ruby gem provides a standard Rails engine, providing out-of-the-box
+defaults for [Shift Commerce](https://shiftcommerce.com/) front-end websites
+written in Rails.
 
-Welcome to your new gem! In this directory, you'll find the files you need to be able to package up your Ruby library into a gem. Put your Ruby code in the file `lib/shift_commerce/rails`. To experiment with that code, run `bin/console` for an interactive prompt.
+# Installing
 
-TODO: Delete this and the text above, and describe your gem
-
-## Installation
-
-Add this line to your application's Gemfile:
+Simply add the gem to your Gemfile:
 
 ```ruby
-gem 'shift_commerce-rails'
+gem 'shiftcommerce-rails'
 ```
-
-And then execute:
-
-    $ bundle
-
-Or install it yourself as:
-
-    $ gem install shift_commerce-rails
-
-## Usage
-
-TODO: Write usage instructions here
-
-## Development
-
-After checking out the repo, run `bin/setup` to install dependencies. Then, run `rake false` to run the tests. You can also run `bin/console` for an interactive prompt that will allow you to experiment.
-
-To install this gem onto your local machine, run `bundle exec rake install`. To release a new version, update the version number in `version.rb`, and then run `bundle exec rake release`, which will create a git tag for the version, push git commits and tags, and push the `.gem` file to [rubygems.org](https://rubygems.org).
-
-## Contributing
-
-Bug reports and pull requests are welcome on GitHub at https://github.com/[USERNAME]/shift_commerce-rails.
-
-
-## License
-
-The gem is available as open source under the terms of the [MIT License](http://opensource.org/licenses/MIT).
-
