--- lib/shift_commerce/rails/version.rb
@@ -1,5 +1,0 @@
-module ShiftCommerce
-  module Rails
-    VERSION = "0.3.1"
-  end
-end
