+++ lib/shift_commerce/templates/shift_variables.erb
