+++ lib/shift_commerce/version.rb
@@ -1,0 +1,3 @@
+module ShiftCommerce
+  VERSION = '0.3.3'.freeze
+end
